rebase and fix merge conflicts
