+++ crates/hotshot/src/tasks/mod.rs
@@ -8,7 +8,11 @@
 
 /// Provides trait to create task states from a `SystemContextHandle`
 pub mod task_state;
-use std::{sync::Arc, time::Duration};
+use crate::{
+    tasks::task_state::CreateTaskState, types::SystemContextHandle, ConsensusApi,
+    ConsensusMetricsValue, ConsensusTaskRegistry, HotShotConfig, HotShotInitializer,
+    MarketplaceConfig, Memberships, NetworkTaskRegistry, SignatureKey, SystemContext, Versions,
+};
 
 use async_broadcast::broadcast;
 use async_compatibility_layer::art::{async_sleep, async_spawn};
@@ -24,11 +28,7 @@ use hotshot_task_impls::rewind::RewindTaskState;
 use hotshot_task_impls::{
     da::DaTaskState,
     events::HotShotEvent,
-    network::{
-        self,
-        test::{ModifierClosure, NetworkEventTaskStateModifier},
-        NetworkEventTaskState, NetworkMessageTaskState,
-    },
+    network::{self, NetworkEventTaskState, NetworkMessageTaskState},
     request::NetworkRequestState,
     response::{run_response_task, NetworkResponseState},
     transactions::TransactionTaskState,
@@ -39,15 +39,15 @@ use hotshot_task_impls::{
 use hotshot_types::message::UpgradeLock;
 use hotshot_types::{
     constants::EVENT_CHANNEL_SIZE,
-    message::Messages,
+    data::QuorumProposal,
+    message::{Messages, Proposal},
     request_response::RequestReceiver,
-    simple_vote::QuorumVote,
     traits::{
         network::ConnectedNetwork,
         node_implementation::{ConsensusTime, NodeImplementation, NodeType},
     },
 };
-use std::fmt::{Debug, Formatter};
+use std::fmt::Debug;
 use std::{collections::HashSet, sync::Arc, time::Duration};
 use vbs::version::StaticVersionType;
 
@@ -736,78 +736,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES> + std::fmt::Debug, V: Version
             }
         }
         vec![event.clone()]
-    }
-}
-
-/// An `EventHandlerState` that modifies view number on the vote of `QuorumVoteSend` event to that of a future view and correctly signs the vote
-pub struct DishonestVoting<TYPES: NodeType> {
-    /// Number added to the original vote's view number
-    pub view_increment: u64,
-    /// A function passed to `NetworkEventTaskStateModifier` to modify `NetworkEventTaskState` behaviour.
-    pub modifier: Arc<ModifierClosure<TYPES>>,
-}
-
-#[async_trait]
-impl<TYPES: NodeType, I: NodeImplementation<TYPES> + std::fmt::Debug, V: Versions>
-    EventTransformerState<TYPES, I, V> for DishonestVoting<TYPES>
-{
-    async fn recv_handler(&mut self, event: &HotShotEvent<TYPES>) -> Vec<HotShotEvent<TYPES>> {
-        vec![event.clone()]
-    }
-
-    async fn send_handler(
-        &mut self,
-        event: &HotShotEvent<TYPES>,
-        public_key: &TYPES::SignatureKey,
-        private_key: &<TYPES::SignatureKey as SignatureKey>::PrivateKey,
-        upgrade_lock: &UpgradeLock<TYPES, V>,
-    ) -> Vec<HotShotEvent<TYPES>> {
-        if let HotShotEvent::QuorumVoteSend(vote) = event {
-            let new_view = vote.view_number + self.view_increment;
-            let spoofed_vote = QuorumVote::<TYPES>::create_signed_vote(
-                vote.data.clone(),
-                new_view,
-                public_key,
-                private_key,
-                upgrade_lock,
-            )
-            .await
-            .context("Failed to sign vote")
-            .unwrap();
-            tracing::debug!("Sending Quorum Vote for view: {new_view:?}");
-            return vec![HotShotEvent::QuorumVoteSend(spoofed_vote)];
-        }
-        vec![event.clone()]
-    }
-
-    fn add_network_event_task(
-        &self,
-        handle: &mut SystemContextHandle<TYPES, I, V>,
-        channel: Arc<<I as NodeImplementation<TYPES>>::Network>,
-        membership: TYPES::Membership,
-        filter: fn(&Arc<HotShotEvent<TYPES>>) -> bool,
-    ) {
-        let network_state: NetworkEventTaskState<_, V, _, _> = NetworkEventTaskState {
-            channel,
-            view: TYPES::Time::genesis(),
-            membership,
-            filter,
-            storage: Arc::clone(&handle.storage()),
-            upgrade_lock: handle.hotshot.upgrade_lock.clone(),
-        };
-        let modified_network_state = NetworkEventTaskStateModifier {
-            network_event_task_state: network_state,
-            modifier: Arc::clone(&self.modifier),
-        };
-        handle.add_task(modified_network_state);
-    }
-}
-
-impl<TYPES: NodeType> Debug for DishonestVoting<TYPES> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("DishonestVoting")
-            .field("view_increment", &self.view_increment)
-            .finish_non_exhaustive()
     }
 }
