+++ crates/testing/tests/tests_1/test_with_failures_2.rs
@@ -72,84 +72,6 @@ cross_tests!(
     }
 );
 
-#[cfg(async_executor_impl = "async-std")]
-cross_tests!(
-    TestName: dishonest_leader,
-    Impls: [MemoryImpl],
-    Types: [TestTypes],
-    Versions: [TestVersions],
-    Ignore: false,
-    Metadata: {
-        let behaviour = Rc::new(|node_id| {
-                let dishonest_leader = DishonestLeader::<TestTypes, MemoryImpl> {
-                    dishonest_at_proposal_numbers: HashSet::from([2, 3]),
-                    validated_proposals: Vec::new(),
-                    total_proposals_from_node: 0,
-                    view_look_back: 1,
-                    _phantom: std::marker::PhantomData
-                };
-                match node_id {
-                    2 => Behaviour::Byzantine(Box::new(dishonest_leader)),
-                    _ => Behaviour::Standard,
-                }
-            });
-
-        let mut metadata = TestDescription {
-            // allow more time to pass in CI
-            completion_task_description: CompletionTaskDescription::TimeBasedCompletionTaskBuilder(
-                                             TimeBasedCompletionTaskDescription {
-                                                 duration: Duration::from_secs(60),
-                                             },
-                                         ),
-            behaviour,
-            ..TestDescription::default()
-        };
-
-        metadata.overall_safety_properties.num_failed_views = 2;
-        metadata.num_nodes_with_stake = 5;
-        metadata.overall_safety_properties.expected_views_to_fail = HashMap::from([
-            (ViewNumber::new(7), false),
-            (ViewNumber::new(12), false)
-        ]);
-        metadata
-    },
-);
-
-cross_tests!(
-    TestName: dishonest_da,
-    Impls: [MemoryImpl, Libp2pImpl, PushCdnImpl],
-    Types: [TestTypes],
-    Versions: [MarketplaceTestVersions],
-    Ignore: false,
-    Metadata: {
-        let behaviour = Rc::new(|node_id| {
-                let dishonest_da = DishonestDa {
-                    dishonest_at_da_cert_sent_numbers: HashSet::from([2]),
-                    total_da_certs_sent_from_node: 0,
-                    total_views_add_to_cert: 4
-                };
-                match node_id {
-                    2 => Behaviour::Byzantine(Box::new(dishonest_da)),
-                    _ => Behaviour::Standard,
-                }
-            });
-
-        let mut metadata = TestDescription {
-            // allow more time to pass in CI
-            completion_task_description: CompletionTaskDescription::TimeBasedCompletionTaskBuilder(
-                                             TimeBasedCompletionTaskDescription {
-                                                 duration: Duration::from_secs(60),
-                                             },
-                                         ),
-            behaviour,
-            ..TestDescription::default()
-        };
-
-        metadata.num_nodes_with_stake = 10;
-        metadata
-    },
-);
-
 cross_tests!(
     TestName: test_with_double_leader_failures,
     Impls: [MemoryImpl, Libp2pImpl, PushCdnImpl],
@@ -193,45 +115,4 @@ cross_tests!(
 
         metadata
     }
-);
-
-cross_tests!(
-    TestName: dishonest_voting,
-    Impls: [MemoryImpl, Libp2pImpl, PushCdnImpl],
-    Types: [TestTypes],
-    Versions: [MarketplaceTestVersions],
-    Ignore: false,
-    Metadata: {
-        let nodes_count: usize = 10;
-        let behaviour = Rc::new(move |node_id| {
-            let dishonest_voting = DishonestVoting {
-                view_increment: nodes_count as u64,
-                modifier: Arc::new(move |_pk, message_kind, transmit_type: &mut TransmitType<TestTypes>, membership: &<TestTypes as NodeType>::Membership| {
-                    if let MessageKind::Consensus(SequencingMessage::General(GeneralConsensusMessage::Vote(vote))) = message_kind {
-                        *transmit_type = TransmitType::Direct(membership.leader(vote.view_number() + 1 - nodes_count as u64));
-                    } else {
-                        {}
-                    }
-                })
-            };
-            match node_id {
-                5 => Behaviour::Byzantine(Box::new(dishonest_voting)),
-                _ => Behaviour::Standard,
-            }
-        });
-
-        let mut metadata = TestDescription {
-            // allow more time to pass in CI
-            completion_task_description: CompletionTaskDescription::TimeBasedCompletionTaskBuilder(
-                                             TimeBasedCompletionTaskDescription {
-                                                 duration: Duration::from_secs(60),
-                                             },
-                                         ),
-            behaviour,
-            ..TestDescription::default()
-        };
-
-        metadata.num_nodes_with_stake = nodes_count;
-        metadata
-    },
 );
