+++ crates/testing/tests/tests_2/byzantine_tests.rs
@@ -11,13 +11,24 @@ use hotshot_macros::cross_tests;
 use hotshot_testing::{
     block_builder::SimpleBuilderImplementation,
     byzantine::byzantine_behaviour::{
-        BadProposalViewDos, DishonestDa, DishonestLeader, DoubleProposeVote, ViewDelay,
+        BadProposalViewDos, DishonestDa, DishonestLeader, DishonestVoting, DoubleProposeVote,
+        ViewDelay,
     },
     completion_task::{CompletionTaskDescription, TimeBasedCompletionTaskDescription},
     test_builder::{Behaviour, TestDescription},
 };
-use hotshot_types::{data::ViewNumber, traits::node_implementation::ConsensusTime};
+use hotshot_types::{
+    data::ViewNumber,
+    message::{GeneralConsensusMessage, MessageKind, SequencingMessage},
+    traits::{
+        election::Membership,
+        network::TransmitType,
+        node_implementation::{ConsensusTime, NodeType},
+    },
+    vote::HasViewNumber,
+};
 use std::rc::Rc;
+use std::sync::Arc;
 cross_tests!(
     TestName: double_propose_vote,
     Impls: [MemoryImpl],
@@ -192,6 +203,47 @@ cross_tests!(
             (ViewNumber::new(21), false),
             (ViewNumber::new(25), false),
         ]);
+        metadata
+    },
+);
+
+cross_tests!(
+    TestName: dishonest_voting,
+    Impls: [MemoryImpl, Libp2pImpl, PushCdnImpl],
+    Types: [TestTypes],
+    Versions: [MarketplaceTestVersions],
+    Ignore: false,
+    Metadata: {
+        let nodes_count: usize = 10;
+        let behaviour = Rc::new(move |node_id| {
+            let dishonest_voting = DishonestVoting {
+                view_increment: nodes_count as u64,
+                modifier: Arc::new(move |_pk, message_kind, transmit_type: &mut TransmitType<TestTypes>, membership: &<TestTypes as NodeType>::Membership| {
+                    if let MessageKind::Consensus(SequencingMessage::General(GeneralConsensusMessage::Vote(vote))) = message_kind {
+                        *transmit_type = TransmitType::Direct(membership.leader(vote.view_number() + 1 - nodes_count as u64));
+                    } else {
+                        {}
+                    }
+                })
+            };
+            match node_id {
+                5 => Behaviour::Byzantine(Box::new(dishonest_voting)),
+                _ => Behaviour::Standard,
+            }
+        });
+
+        let mut metadata = TestDescription {
+            // allow more time to pass in CI
+            completion_task_description: CompletionTaskDescription::TimeBasedCompletionTaskBuilder(
+                                             TimeBasedCompletionTaskDescription {
+                                                 duration: Duration::from_secs(60),
+                                             },
+                                         ),
+            behaviour,
+            ..TestDescription::default()
+        };
+
+        metadata.num_nodes_with_stake = nodes_count;
         metadata
     },
 );
